requirements.txt
# Python Dependencies for Slug Sight Ground Station
# Install with: pip install -r requirements.txt

# Core dependencies
pyserial>=3.5          # Serial communication with ground Feather M4
pyyaml>=6.0            # Configuration file parsing
pandas>=1.5.0          # Data handling and CSV export

# Flask web server
Flask>=2.3.0           # Web framework
Flask-Cors>=4.0.0      # CORS support (optional)

# Visualization (optional)
matplotlib>=3.6.0      # Plotting telemetry data
numpy>=1.23.0          # Numerical operations

# Development/Testing
pytest>=7.2.0          # Unit testing
pytest-cov>=4.0.0      # Test coverage

# Logging and utilities
colorama>=0.4.6        # Colored terminal output
python-dateutil>=2.8.2 # Date/time utilities
